requirements.txt
cffi==1.15.0
cryptography==35.0.0
decorator==5.1.0
psycopg2-binary==2.9.2
pycparser==2.21
six==1.16.0
tabulate==0.8.9
